# apps/docs/static/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/calendar/google/connect": {
      "get": {
        "operationId": "CalendarController_connect_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Démarre l'autorisation Google Calendar (navigation du navigateur, session requise)",
        "tags": [
          "calendar"
        ]
      }
    },
    "/v1/calendar/google/callback": {
      "get": {
        "operationId": "CalendarController_callback_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "summary": "Retour de Google après l'autorisation d'accès à l'agenda (redirection)",
        "tags": [
          "calendar"
        ],
        "security": []
      }
    },
    "/v1/households/{householdId}/calendar": {
      "get": {
        "operationId": "CalendarController_status_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarStatusDto"
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "État de la synchronisation Google Agenda",
        "tags": [
          "calendar"
        ]
      }
    },
    "/v1/households/{householdId}/calendar/available": {
      "get": {
        "operationId": "CalendarController_available_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AvailableCalendarDto"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Agendas Google disponibles",
        "tags": [
          "calendar"
        ]
      }
    },
    "/v1/households/{householdId}/calendar/link": {
      "put": {
        "operationId": "CalendarController_link_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkCalendarInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarStatusDto"
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Choisir l'agenda Google synchronisé",
        "tags": [
          "calendar"
        ]
      },
      "delete": {
        "operationId": "CalendarController_unlink_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "removeEvents",
            "required": false,
            "in": "query",
            "schema": {
              "default": "true",
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Aucun contenu"
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Arrêter la synchronisation du foyer",
        "tags": [
          "calendar"
        ]
      }
    },
    "/v1/households/{householdId}/calendar/sync": {
      "post": {
        "description": "« Synchroniser maintenant » : réconciliation + balayage en arrière-plan.",
        "operationId": "CalendarController_syncNow_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Succès"
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "« Synchroniser maintenant » ",
        "tags": [
          "calendar"
        ]
      }
    },
    "/v1/me/google-calendar": {
      "delete": {
        "operationId": "CalendarController_disconnect_v1",
        "parameters": [
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Aucun contenu"
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Révoquer l'accès à Google Agenda",
        "tags": [
          "calendar"
        ]
      }
    },
...
  },
  "info": {
    "title": "Agenda G & N — API",
    "description": "API REST de l'application **Agenda G & N** (web et Android). Toutes les routes sont versionnées\nsous `/v1`. Guide complet (authentification, erreurs, idempotence, temps réel) : page\n*API → Guide* de la documentation.\n\n- **Authentification** : `Authorization: Bearer <accessToken>` (Android, en-tête\n  `X-Client: mobile` à la connexion pour recevoir les jetons dans le corps), ou cookies\n  `httpOnly` posés par `/v1/auth/login` (web).\n- **CSRF** : toute requête qui modifie (POST, PATCH, PUT, DELETE) doit porter\n  `x-requested-with: agenda-gn`.\n- **Erreurs** : toujours `{ \"error\": { \"code\", \"message\", \"details\"? } }` ; le `code` est\n  stable (voir le schéma `ApiError`), le `message` n'est destiné qu'aux développeurs.\n- **Idempotence** : les créations de tâches acceptent `Idempotency-Key` (UUID) pour être\n  rejouées sans doublon depuis une file hors ligne.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Même origine que le site (le web relaie /v1/* vers l’API)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gn_at"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_FAILED",
                  "UNAUTHENTICATED",
                  "INVALID_CREDENTIALS",
                  "EMAIL_ALREADY_USED",
                  "REGISTRATION_CLOSED",
                  "ACCOUNT_DISABLED",
                  "SESSION_EXPIRED",
                  "CSRF_REJECTED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "HOUSEHOLD_NOT_FOUND",
                  "INVITATION_INVALID",
                  "RESET_TOKEN_INVALID",
                  "GOOGLE_NOT_CONFIGURED",
                  "GOOGLE_FAILED",
                  "GOOGLE_EMAIL_EXISTS",
                  "GOOGLE_ALREADY_LINKED",
                  "CURRENT_PASSWORD_INVALID",
                  "PASSWORD_REQUIRED",
                  "CONFIRMATION_REQUIRED",
                  "CALENDAR_NOT_CONFIGURED",
                  "CALENDAR_NOT_CONNECTED",
                  "CALENDAR_READ_ONLY",
                  "CALENDAR_NOT_FOUND",
                  "GOOGLE_REVOKED",
                  "GOOGLE_FLOW_EXPIRED",
                  "GOOGLE_DENIED",
                  "GOOGLE_SCOPES_MISSING",
                  "ALREADY_MEMBER",
                  "VERSION_CONFLICT",
                  "IDEMPOTENCY_KEY_INVALID",
                  "IDEMPOTENCY_KEY_REUSED",
                  "IDEMPOTENCY_IN_PROGRESS",
                  "CATEGORY_NAME_TAKEN",
                  "TASK_TITLE_REQUIRED",
                  "RATE_LIMITED",
                  "ATTACHMENT_TOO_LARGE",
                  "ATTACHMENTS_QUOTA",
                  "INTERNAL"
                ]
              },
              "message": {
                "type": "string",
                "description": "Message technique (anglais), jamais affiché."
              },
              "details": {
                "description": "Détail éventuel (champs invalides pour VALIDATION_FAILED)."
              }
            }
          }
        }
      },
...
      "AvailableCalendarDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "accessRole": {
            "type": "string"
          },
          "primary": {
            "type": "boolean"
          },
          "writable": {
            "type": "boolean"
          },
          "color": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "summary",
          "accessRole",
          "primary",
          "writable",
          "color"
        ],
        "additionalProperties": false
      },
...
      "CalendarStatusDto": {
        "type": "object",
        "properties": {
          "configured": {
            "type": "boolean"
          },
          "connection": {
            "nullable": true,
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "email": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "ACTIVE",
                  "REVOKED",
                  "ERROR"
                ]
              }
            },
            "required": [
              "id",
              "email",
              "status"
            ],
            "additionalProperties": false
          },
          "link": {
            "nullable": true,
            "type": "object",
            "properties": {
              "calendarId": {
                "type": "string"
              },
              "summary": {
                "type": "string"
              },
              "accessRole": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "ACTIVE",
                  "INVALID"
                ]
              },
              "errorCode": {
                "nullable": true,
                "type": "string",
                "enum": [
                  "CALENDAR_READ_ONLY",
                  "CALENDAR_NOT_FOUND",
                  "GOOGLE_REVOKED",
                  "CALENDAR_RATE_LIMITED",
                  "CALENDAR_UNAVAILABLE"
                ]
              },
              "connectedByMe": {
                "type": "boolean"
              },
              "connectedBy": {
                "nullable": true,
                "type": "string"
              },
              "connectionEmail": {
                "type": "string"
              },
              "lastReconciledAt": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "calendarId",
              "summary",
              "accessRole",
              "status",
              "errorCode",
              "connectedByMe",
              "connectedBy",
              "connectionEmail",
              "lastReconciledAt"
            ],
            "additionalProperties": false
          },
          "stats": {
            "type": "object",
            "properties": {
              "synced": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "pending": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "errors": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "synced",
              "pending",
              "errors"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "configured",
          "connection",
          "link",
          "stats"
        ],
        "additionalProperties": false
      },
...
      "LinkCalendarInput": {
        "type": "object",
        "properties": {
          "calendarId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          }
        },
        "required": [
          "calendarId"
        ]
      },
...
    }
  }
}